// SGUEES-SPA/src/app/pages/Shop/com-cotizacion/models/com-cotizacion.ts
export interface ComCotizacion {
	CORR_EMPRESA: number;
	ANIO_PERIODO: number;
	CORR_COTIZACION: number;
  NUMERO_COTIZACION: string;
	FECHA_COTIZACION: Date;
	CORR_PROVEEDOR: number;
	NOMBRE_PROVEEDOR: string;
	USUARIO_COTIZA: string;
	OBSERVACIONES: string;
	OBSERVACIONES_PROVEEDOR: string;
	PLAZO_ENTREGA: string;
	ESTADO_COTIZACION: string;
	NOMBRE_ESTADO_COTIZACION: string;
	ANIO_PERIODO_SOLI_COTI: number;
	CORR_SOLI_COTIZACION: number;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
  NUMERO_SOLI_COTIZACION: string;
	NUMERO_SOLI_COMPRA: string;
	CORR_CONDICION_PAGO: number;
	NOMBRE_CONDICION_PAGO: string;
}